refactor: use floats for Splines

+++ src/main/java/io/github/foundationgames/splinecart/util/InterpolationResult.java
@@ -1,10 +1,10 @@
 package io.github.foundationgames.splinecart.util;
 
-import org.joml.Matrix3d;
-import org.joml.Vector3d;
+import org.joml.Matrix3f;
+import org.joml.Vector3f;
 
-public record InterpolationResult(Vector3d translation, Matrix3d basis, Vector3d gradient  /* Change in position per change in spline progress */) {
+public record InterpolationResult(Vector3f translation, Matrix3f basis, Vector3f gradient  /* Change in position per change in spline progress */) {
     public InterpolationResult() {
-        this(new Vector3d(), new Matrix3d(), new Vector3d());
+        this(new Vector3f(), new Matrix3f(), new Vector3f());
     }
 }
